requirements.txt
flask==2.3.3
openai==0.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.32
werkzeug==2.3.7
